auth_api/python/src/requirements.txt
Flask~=2.2.2
python-dotenv~=0.21.1
PyJWT~=2.6.0
pymysql~=1.0.2